+++ changelog.md
@@ -24,6 +24,7 @@
 ### Improvements
 
 - #214 Add backend menu entry for each Rule Builder scope.
+- Improvements in the processing of recurring payments.
 
 ### Bugfixes